% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_unidadvecinal_fct_localizacion.R
\name{get_uv_with_address}
\alias{get_uv_with_address}
\title{Get Unidad Vecinal With Address}
\usage{
get_uv_with_address(direccion, comuna = NULL, region = NULL)
}
\value{
Unidad Vecinal
}
\description{
Get Unidad Vecinal With Address
}
